Set up proof of concept for drag and drop
Created a simplified view of the music player with a drop target that autoloads and plays the selected song. I think I'll try to start simple and add complexity when it's necessary. As such, I'm also setting the data into a component state instead of building the store right away.