Fixed display bug, added redux store and testing
Fixed a display bug that caused the now playing bar to become obscured when the queue fill up and begins to overflow it's container. The issue was fixed by adding 'flex-shrink: 0;' to the other components in order to prevent them from resizing to accomodate the negative space (that's what 'overflow-y: auto;' is for in the queue list anyways.)

Also, I switched from holding the queue in two separate state variables representing the current song and the other songs to a single queue list in a redux store. This will help with complexity as the data structure of the application grows, though it does not produce immediate benefits yet.

Finally, I began using that test folder I added in order to test the new Redux store I implemented using mocha and chai. They are nice since they produce tests in a very human readable format.