Added playlist sorting + search bar
Added a search bar to ther playlist area to filter and display only songs that contain the substring specified (case insensitive). Also, I removed a console.log statement (how did that little thing get in there?), modified the css again slightly to increase the size of the song items for readability, and fixed the issue where playling the same song twice in a row in the queue wouldn't autoplay the next song.